Panda and Chain Reaction
Panda has become a scientist recently. In his laboratory, there are infinite number of chambers where a chamber number K is connected to a chamber number K-1.

The numbering of chambers start from 0. Initially, X number of particles are present in the chamber number 0. The number of particles present in chamber K is K times the number of particles present in chamber K-1. You are supposed to help Panda in finding out the number of particles in a given chamber number N.

Note: The number of particles in chamber K cannot be calculated until the number of particles in chamber K-1 are calculated.

Constraints:

1 <= T <= 105

Subtask 1: (20 points)
1 <= N, X <= 10**5

Subtask 2: (80 points)
1 <= N, X <= 10**18