Riya's Birthday
Madhav went to Riya's Birthday Party. He was a geek so he had no idea regarding which gift she'l like.
So he took an array of integers with him. The array followed a particular order. First element of array is 1.
Second element of array is 6.
Other elements of the array are two less than the mean of the number preceding and succeeding it.
As it is obvious, Riya felt that this idea was stupid and hence she wanted to punish Madhav.
She decided to ask Madhav the nth number of the array. If he tells the wrong answer, she would slap him.
Help Madhav to escape from this embarrassing situation.